[up] RTAlightcurve.py :: updated recap print, changed some prints --left essentials only-- and fixed the +2h thing by controlling with is_detection list of boolean if it's first none detection.

+++ RTAlightcurve.py
@@ -28,7 +28,7 @@ tmin = 30  # slewing time (s)
 tmax = []
 for i in range(len(texp)):
   tmax.append(tmin + texp[i])
-ttotal = 300 # 2e4  # simulation total time (s)
+ttotal = 600 # 14400 (4h) simulation total time (s)
 elow = 0.03  # simulation minimum energy (TeV)
 ehigh = 150.0  # simulation maximum energy (TeV)
 emin = 0.03  # selection minimum energy (TeV)
@@ -42,27 +42,16 @@ ts_threshold = 25  # TS threshold for reliable detection
 reduce_flux = None  # flux will be devided by factor reduce_flux, if nominal then set to None ---!
 
 # conditions control ---!
-checks = True  # prints checks info ---!
-if_ebl = False  # uses the EBL absorbed template ---!
+checks = False  # prints checks info ---!
+if_ebl = True  # uses the EBL absorbed template ---!
 if_cut = False  # adds a cut-off parameter to the source model ---!
 ebl_fits = False  # generate the EBL absorbed template ---!
 extract_spec = True  # generates spectral tables and obs definition models ---!
 irf_degrade = False  # use degraded irf ---!
 src_sort = False  # sorts scandidates from highest TS to lowest ---!
-skip_exist = False  # if an output already exists it skips the step ---!
+skip_exist = True  # if an output already exists it skips the step ---!
 debug = False  # prints logfiles on terminal ---!
 if_log = True  # saves logfiles ---!
-
-# recap and dof ---!
-dof, m2, m1 = getDof()
-print('!!! *** !!! dof = ', m2, ' - ', m1, ' = ', dof)
-print('!!! *** !!! EBL ABSORPTION:', if_ebl)
-print('!!! *** !!! MODEL CUTOFF:', if_cut)
-print('!!! *** !!! IRF DEGRADATION:', irf_degrade)
-print('!!! *** !!! TS SORT:', src_sort)
-print('!!! *** !!! sim energy range: [', elow, ', ', ehigh, '] (TeV)')
-print('!!! *** !!! selection energy range: [', emin, ', ', emax, '] (TeV)')
-print('!!! *** !!! roi: ', roi, ' (deg)')
 
 # path configuration ---!
 cfg = xmlConfig()
@@ -82,7 +71,22 @@ offmax = (-1.475, -1.370)  # off-axis RA/DEC (deg)
 pointing = (true_coord[0] + offmax[0], true_coord[1] + offmax[1])  # pointing direction RA/DEC (deg)
 # true_coord, pointing, offmax = getPointing(None, p.getWorkingDir()+nominal_template)
 # pointing with off-axis equal to max prob GW ---!
-print(true_coord, pointing, offmax) if checks is True else None
+print(true_coord, pointing, offmax) if checks else None
+
+# recap and dof ---!
+dof, m2, m1 = getDof()
+print('!!! *** !!! dof = ', m2, ' - ', m1, ' = ', dof)
+print('!!! *** !!! EBL ABSORPTION:', if_ebl)
+print('!!! *** !!! MODEL CUTOFF:', if_cut)
+print('!!! *** !!! IRF DEGRADATION:', irf_degrade)
+print('!!! *** !!! nominal caldb:', caldb)
+print('!!! *** !!! irf:', irf)
+print('!!! *** !!! TS SORT:', src_sort)
+print('!!! *** !!! FLUX REDUCED factor:', reduce_flux)
+print('!!! *** !!! sim energy range: [', elow, ', ', ehigh, '] (TeV)')
+print('!!! *** !!! selection energy range: [', emin, ', ', emax, '] (TeV)')
+print('!!! *** !!! roi: ', roi, ' (deg)')
+print('!!! *** !!! pointing:', pointing, ' (deg)')
 
 # --------------------------------- INITIALIZE --------------------------------- !!!
 
@@ -112,25 +116,25 @@ if ebl_fits:
 # assign template ---!
 if if_ebl:
   template = p.getWorkingDir() + ebl_template
-  print('with EBL') if checks is True else None
+  print('with EBL') if checks else None
 else :
   template = p.getWorkingDir() + nominal_template
-  print('w/o EBL') if checks is True else None
+  print('w/o EBL') if checks else None
 tObj.if_ebl = if_ebl
 tObj.template = template
-print('!!! check ---- template=', tObj.template) if checks is True else None
+print('!!! check ---- template=', tObj.template) if checks else None
 # load template ---!
 tObj.extract_spec = extract_spec
 tbin_stop = tObj.loadTemplate()
-print('!!! check ---- tbin_stop=', tbin_stop) if checks is True else None
-print('!!! check ---- caldb:', tObj.caldb)
+print('!!! check ---- tbin_stop=', tbin_stop) if checks else None
+print('!!! check ---- caldb:', tObj.caldb) if checks else None
 
 # --------------------------------- REDUCE TEMPLATE FLUX  --------------------------------- !!!
 
 if reduce_flux != None:
   tObj.factor = reduce_flux
   tObj.makeFainter()
-  print('!!! check ---- reduce flux by factor %s' %str(reduce_flux)) if checks is True else None
+  print('!!! check ---- reduce flux by factor %s' %str(reduce_flux)) if checks else None
 
 # --------------------------------- 1° LOOP :: trials  --------------------------------- !!!
 
@@ -138,8 +142,8 @@ for k in range(trials):
   count += 1
   tObj.seed = count
   clocking = 0  # simulate flowing time (subsequent temporal windows of 1s)
-  print('\n\n!!! ************ STARTING TRIAL %d ************ !!!\n\n' %count) if checks is True else None
-  print('!!! check ---- seed=', tObj.seed) if checks is True else None
+  print('\n\n!!! ************ STARTING TRIAL %d ************ !!!\n\n' %count) if checks else None
+  print('!!! check ---- seed=', tObj.seed) if checks else None
   # attach ID to fileroot ---!
   if if_ebl:
     f = fileroot + 'ebl%06d' % (count)
@@ -147,7 +151,7 @@ for k in range(trials):
     f = fileroot + 'sim%06d' % (count)
   if irf_degrade:
     f += 'irf'
-  print('!!! check ---- file=', f) if checks is True else None
+  print('!!! check ---- file=', f) if checks else None
 
   # --------------------------------- SIMULATION --------------------------------- !!!
 
@@ -167,7 +171,7 @@ for k in range(trials):
       tObj.model = tObj.model.replace('_tbin', '_flux%s_tbin' %str(reduce_flux))
       event = event .replace('_tbin', '_flux%s_tbin' %str(reduce_flux))
     event_bins.append(event)
-    if not skip_exist:
+    if not skip_exist or not os.path.isfile(event):
       if os.path.isfile(event):
         os.remove(event)
       tObj.output = event
@@ -177,7 +181,7 @@ for k in range(trials):
   event_list = p.getSimDir() + 'obs_%s.xml' % f
   if reduce_flux != None:
     event_list = event_list.replace('obs_', 'obs_flux%s_' %str(reduce_flux))
-  if not skip_exist:
+  if not skip_exist or not os.path.isfile(event_list):
     if os.path.isfile(event_list):
       os.remove(event_list)
     tObj.input = event
@@ -189,13 +193,16 @@ for k in range(trials):
   # --------------------------------- SELECTION --------------------------------- !!!
 
   tObj.e = [emin, emax]
-  twindows = [ttotal/texp[i] for i in range(len(texp))]  # number of temporal windows per exposure time in total time
-  tlast = [ttotal for i in range(len(texp))]
+  twindows = [ttotal/texp[i] for i in range(len(texp))]  # number of temporal windows per exposure time in total time ---!
+  tlast = [ttotal for i in range(len(texp))]  # maximum observation time from last detection (not exceeding ttotal) ---!
+  is_detection = [True for i in range(len(texp))]  # controls which avoid forwarding of tlast for subsequent non-detections ---!
   # looping for all lightcurve second by second ---!
   for j in range(int(max(twindows))):
     clocking += 1  # passing time second by second ---!
+    print(clocking, 'j loop', tlast, is_detection)
     # check tlast, if globally reached then stop current trial ---!
     if clocking > max(tlast):
+      print('end analysis trial', count)
       break
     current_twindows = []
     for i in range(len(texp)):
@@ -203,7 +210,10 @@ for k in range(trials):
     # looping for all the texp for which the tbin analysis needs to be computed ---!
     for i in range(len(current_twindows)):
       # check tlast, if locally reached then skip current bin ---!
-      if clocking > tlast[i]:
+      index = texp.index(current_twindows[i])
+      print('i loop', tlast, is_detection, index)
+      if clocking > tlast[index]:
+        print('skip analysis texp', texp[index])
         continue
       tbin = clocking/current_twindows[i] # temporal bin number of this analysis
       # data file init and check to avoid doubles ---!
@@ -215,11 +225,12 @@ for k in range(trials):
         tObj.t = [tmin, tmax[i]]
       else:
         tObj.t = [tmin+clocking, tmax[i]+clocking]
-      print(tmin, tmax[i]) if (clocking in texp) else print(tmin+clocking, tmax[i]+clocking)
+      # select events ---!
       event_selected = event_list.replace(p.getSimDir(), p.getSelectDir()).replace('obs_', 'texp%ds_tbin%d_' %(texp[i], tbin))
       prefix = p.getSelectDir() + 'texp%ds_tbin%d_' %(texp[i], tbin)
-      if os.path.isfile(event_selected):
-        os.remove(event_selected)
+      if not skip_exist or not os.path.isfile(event_selected):
+        if os.path.isfile(event_selected):
+          os.remove(event_selected)
       tObj.input = event_list
       tObj.output = event_selected
       tObj.eventSelect(prefix=prefix)
@@ -227,8 +238,9 @@ for k in range(trials):
       # --------------------------------- SKYMAP --------------------------------- !!!
 
       skymap = event_selected.replace(p.getSelectDir(), p.getDetDir()).replace('.xml', '_skymap.fits')
-      if os.path.isfile(skymap):
-        os.remove(skymap)
+      if not skip_exist or not os.path.isfile(skymap):
+        if os.path.isfile(skymap):
+          os.remove(skymap)
       tObj.input = event_selected
       tObj.output = skymap
       tObj.eventSkymap(wbin=wbin)
@@ -238,8 +250,9 @@ for k in range(trials):
       tObj.corr_rad = corr_rad
       tObj.max_src = max_src
       detectionXml = skymap.replace('_skymap.fits', '_det%dsgm.xml' %sigma)
-      if os.path.isfile(detectionXml):
-        os.remove(detectionXml)
+      if not skip_exist or not os.path.isfile(detectionXml):
+        if os.path.isfile(detectionXml):
+          os.remove(detectionXml)
       tObj.input = skymap
       tObj.output = detectionXml
       tObj.runDetection()
@@ -247,13 +260,14 @@ for k in range(trials):
       detObj.sigma = sigma
       detObj.if_cut = if_cut
       detObj.modXml()
+      detObj.prmsFreeFix()
 
       # --------------------------------- MAX LIKELIHOOD --------------------------------- !!!
 
-      detObj.prmsFreeFix()
       likeXml = detectionXml.replace('_det%dsgm' % tObj.sigma, '_like%dsgm' % tObj.sigma)
-      if os.path.isfile(likeXml):
-        os.remove(likeXml)
+      if not skip_exist or not os.path.isfile(likeXml):
+        if os.path.isfile(likeXml):
+          os.remove(likeXml)
       tObj.input = event_selected
       tObj.model = detectionXml
       tObj.output = likeXml
@@ -261,15 +275,14 @@ for k in range(trials):
       likeObj = ManageXml(likeXml)
       if src_sort:
         highest_ts_src = likeObj.sortSrcTs()[0]
+        print('!!! check ---- highest TS: ', highest_ts_src) if checks else None
       else:
         highest_ts_src = None
-        print('!!! check ---- highest TS: ', highest_ts_src) if checks is True else None
 
       # --------------------------------- DETECTION RA & DEC --------------------------------- !!!
 
-      pos, ra_det, dec_det = ([] for j in range(3))
+      pos, ra_det, dec_det = ([] for n in range(3))
       pos.append(detObj.loadRaDec(highest=highest_ts_src))
-      print('!!! check ---- coords:', pos[0]) if checks is True else None
       ra_det.append(pos[0][0][0]) if len(pos[0][0]) > 0 else ra_det.append(np.nan)
       dec_det.append(pos[0][1][0]) if len(pos[0][0]) > 0 else dec_det.append(np.nan)
       Ndet = len(pos[0][0])
@@ -280,11 +293,12 @@ for k in range(trials):
 
       # --------------------------------- BEST FIT TSV --------------------------------- !!!
 
-      ts_list, ts = ([] for j in range(2))
+      ts_list, ts = ([] for n in range(2))
       ts_list.append(likeObj.loadTs()) if Ndet > 0 else ts_list.append([np.nan])
 
       # only first elem ---!
       ts.append(ts_list[0][0])
+      print(ts[0])
 
       # --------------------------------- Nsrc FOR TSV THRESHOLD --------------------------------- !!!
 
@@ -298,16 +312,19 @@ for k in range(trials):
 
       # --------------------------------- +2h FROM LAST DETECTION --------------------------------- !!!
 
-      if Nsrc == 0:
+      if (float(ts[0]) < ts_threshold or float(ts[0]) == np.nan) and is_detection[index]:
+        is_detection[index] = False
         # add 2hrs of obs time ---!
-        tlast[i] = clocking+7200
+        tlast[index] = clocking+5  # 7200 # +2h ---!
+        print('+2h tlast = ', tlast[index], ' with texp = ', texp[index])
         # only 4hrs of simulation avialable, if tlast exceeds them then reset to ttotal ---!
-        if tlast[i] > ttotal:
-          tlast[i] = ttotal
+        if tlast[index] > ttotal:
+          tlast[index] = ttotal
+          print('reset tlast = ', tlast[index], ' with texp = ', texp[index])
 
       # --------------------------------- BEST FIT RA & DEC --------------------------------- !!!
 
-      ra_list, ra_fit, dec_list, dec_fit = ([] for j in range(4))
+      ra_list, ra_fit, dec_list, dec_fit = ([] for n in range(4))
       coord = likeObj.loadRaDec() if Ndet > 0 else None
       ra_list.append(coord[0]) if Ndet > 0 else ra_list.append([np.nan])
       dec_list.append(coord[1]) if Ndet > 0 else dec_list.append([np.nan])
@@ -318,7 +335,7 @@ for k in range(trials):
 
       # --------------------------------- BEST FIT SPECTRAL --------------------------------- !!!
 
-      pref_list, pref, index_list, index, pivot_list, pivot = ([] for j in range(6))
+      pref_list, pref, index_list, index, pivot_list, pivot = ([] for n in range(6))
       likeObj.if_cut = if_cut
       spectral = likeObj.loadSpectral()
       index_list.append(spectral[0]) if Ndet > 0 else index_list.append([np.nan])
@@ -332,7 +349,7 @@ for k in range(trials):
 
       # eventually cutoff ---!
       if if_cut:
-        cutoff_list, cutoff = ([] for j in range(2))
+        cutoff_list, cutoff = ([] for n in range(2))
         cutoff_list.append(spectral[3]) if Ndet > 0 else cutoff_list.append([np.nan])
         cutoff.append(cutoff_list[0][0])
 
@@ -359,8 +376,8 @@ for k in range(trials):
 
       row = []
       if checks:
-        print('\n\n!!! ---------- check trial:', count)
-        print('!!! ----- check texp:', texp[i])
+        print('!!! ---------- check trial:', count)
+        print('!!! ----- check texp:', texp[i], 's between: [', tObj.t[0], ', ', tObj.t[1], ' ] s')
         print('!!! *** check Ndet:', Ndet)
         print('!!! *** check Nsrc:', Nsrc)
         print('!!! *** check ra_det:', ra_det[0])
@@ -370,6 +387,7 @@ for k in range(trials):
         print('!!! *** check flux_ph:', flux_ph[0])
         # print('!!! *** check flux_en:', flux_en[i][0])
         print('!!! *** check ts:', ts[0])
+        print('!!! *** ---------------------------')
 
       row.append([ID, texp[i], sigma, Ndet, Nsrc, ra_det[0], dec_det[0], ra_fit[0], dec_fit[0],
                   flux_ph[0], flux_en[0], ts[0]])
@@ -387,8 +405,8 @@ for k in range(trials):
 
   # --------------------------------- CLEAR SPACE --------------------------------- !!!
 
-  print('!!! check ---- ', count, ') trial done...') if checks is True else None
-  if count not in [1,2,3,4]:
+  print('!!! check ---- ', count, ') trial done...') if checks else None
+  if count != 1:
     os.system('rm ' + p.getSimDir() + '*run*%06d*' % count)
     os.system('rm ' + p.getSelectDir() + '*run*%06d*' % count)
     os.system('rm ' + p.getDetDir() + '*run*%06d*' % count)
